+++ src/core/io/svg/svg_parser.cpp
@@ -96,6 +96,17 @@ public:
             }
         }
 
+        for ( const auto& link_node : ItemCountRange(dom.elementsByTagName("link")) )
+        {
+            auto link = link_node.toElement();
+            if ( link.attribute("rel") == "stylesheet" )
+            {
+                QString url = link.attribute("href");
+                if ( !url.isEmpty() )
+                    document->add_pending_asset(url);
+            }
+        }
+
         parse_css();
         parse_defs();
 
@@ -142,6 +153,9 @@ public:
                 if ( child.isText() || child.isCDATASection() )
                     data += child.toCharacterData().data();
             }
+
+            if ( data.contains("@font-face") )
+                document->add_pending_asset(data.toUtf8());
 
             parser.parse(data);
         }
@@ -1411,7 +1425,8 @@ public:
         qfont.setWeight(WeightConverter::convert(style.weight, WeightConverter::css, WeightConverter::qt));
         qfont.setStyle(style.style);
         QFontDatabase db;
-        font->style.set(db.styleString(qfont));
+        QString style_string = db.styleString(qfont);
+        font->style.set(style_string);
     }
 
     QPointF parse_text_element(const ParseFuncArgs& args, const TextStyle& parent_style)