+++ src/core/model/assets/pending_asset.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <QUrl>
+#include <QByteArray>
+
+
+namespace glaxnimate::model {
+
+struct PendingAsset
+{
+    int id = 0;
+    QUrl url;
+    QByteArray data;
+    bool loaded = false;
+};
+
+} // namespace glaxnimate::model